Moved pipes to sub-directory. Started work on fixing tests for new back-pressure system.
