+++ src/NexNet/Internals/NexusSession.cs
@@ -11,6 +11,7 @@ using System.Collections.Concurrent;
 using Pipelines.Sockets.Unofficial.Threading;
 using Pipelines.Sockets.Unofficial.Buffers;
 using System.Runtime.CompilerServices;
+using NexNet.Internals.Pipes;
 
 namespace NexNet.Internals;
